Generate interactive KM plot by endpoints and subgroups

Generate interactive KM plot by endpoints and subgroups

kmcurvely(
  meta = meta_tte_example(),
  population = "apat",
  observation = "efficacy_population",
  endpoint = "pfs;ttde",
  subgroup = "male;female",
  x_label = NULL,
  y_label = "Survival rate",
  color = NULL,
  time_unit = c("days", "weeks", "months", "years")
)

Arguments

meta

A metadata object that contains the mappings for the variables of interest.

population

A string indicating the population to use (e.g., "apat").

observation

A string indicating the observation population (e.g., "efficacy_population").

endpoint

A semicolon-separated string of endpoints to be analyzed (e.g., "pfs;ttde").

subgroup

A semicolon-separated string of subgroups to filter by (e.g., "male;female").

x_label

A string for the x-axis label (default is "Time in days/weeks/months/years").

y_label

A string for the y-axis label (default is "Survival rate").

color

A vector of colors for the plot (if NULL, defaults to a preset color palette).

time_unit

A string specifying the unit of time; must be one of "days", "weeks", "months", or "years". Default is "days".

Value

An interactive KM plot

Examples

# Example usage of kmcurvely function

meta <- meta_tte_example()
kmcurvely(
  meta = meta,
  population = "apat",
  observation = "efficacy_population",
  endpoint = "pfs;ttde",
  subgroup = "male;female"
)
#> Joining with `by = join_by(USUBJID, TRT01P, SEX)`
#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if
#> `.name_repair` is omitted as of tibble 2.0.0.
#> ℹ Using compatibility `.name_repair`.
#> ℹ The deprecated feature was likely used in the kmcurvely package.
#>   Please report the issue to the authors.
#> Joining with `by = join_by(time)`
#> Joining with `by = join_by(time)`
#> Joining with `by = join_by(time)`
#> Joining with `by = join_by(time)`
#> Joining with `by = join_by(time)`
#> Joining with `by = join_by(time)`
#> Warning: Ignoring unknown aesthetics: text
#> Warning: The group argument has been deprecated. Use group_by() or split instead.
#> Warning: Removed 6 rows containing missing values or values outside the scale range
#> (`geom_point()`).
#> Warning: Removed 2 rows containing missing values or values outside the scale range
#> (`geom_point()`).
#> Warning: Removed 6 rows containing missing values or values outside the scale range
#> (`geom_point()`).

  

    

      

        Endpoint
        

          

      

      

        Subgroup
        

          

      

    

    

      

      
      Number of participants at risk
      

      

  

Download All Plots